flatten renderer modules
