+++ src/backend/renderer/mod.rs
@@ -38,7 +38,10 @@ pub mod multigpu;
 pub mod utils;
 
 #[cfg(feature = "wayland_frontend")]
-pub mod output;
+pub mod element;
+
+#[cfg(feature = "wayland_frontend")]
+pub mod damage;
 
 #[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
 /// Texture filtering methods
