+++ src/desktop/space/window.rs
@@ -1,5 +1,11 @@
 use crate::{
-    backend::renderer::{ImportAll, Renderer, Texture},
+    backend::renderer::{
+        element::{
+            surface::{render_elements_from_surface_tree, WaylandSurfaceRenderElement},
+            RenderElement,
+        },
+        ImportAll, Renderer, Texture,
+    },
     desktop::{window::Window, PopupManager},
     utils::{Logical, Physical, Point, Rectangle, Scale},
 };
@@ -48,8 +54,7 @@ impl<R, E> SpaceElement<R, E> for Window
 where
     R: Renderer + ImportAll,
     <R as Renderer>::TextureId: Texture + 'static,
-    E: crate::backend::renderer::output::element::RenderElement<R>
-        + From<crate::backend::renderer::output::element::surface::WaylandSurfaceRenderElement>,
+    E: RenderElement<R> + From<WaylandSurfaceRenderElement>,
 {
     fn location(&self, space_id: usize) -> Point<i32, Logical> {
         window_loc(self, &space_id) - self.geometry().loc
@@ -74,20 +79,12 @@ where
                     .to_physical(scale)
                     .to_i32_round();
 
-                crate::backend::renderer::output::element::surface::surfaces_from_surface_tree(
-                    popup.wl_surface(),
-                    location + offset,
-                    scale,
-                )
+                render_elements_from_surface_tree(popup.wl_surface(), location + offset, scale)
             });
 
         render_elements.extend(popup_render_elements);
 
-        render_elements.extend(
-            crate::backend::renderer::output::element::surface::surfaces_from_surface_tree(
-                surface, location, scale,
-            ),
-        );
+        render_elements.extend(render_elements_from_surface_tree(surface, location, scale));
 
         render_elements
     }
